feat: added the number field component
